requirements.txt
requests>=2.25.1
feedparser
newspaper3k
GoogleNews
beautifulsoup4
lxml
pandas
transformers
lxml_html_clean
cssselect
pillow
PyYAML
typing-extensions
tqdm
tldextract
tinysegmenter
tzdata
tzlocal
google-generativeai
torch
python-dotenv
